Initial Commit - GUI
GUI (Login, Register, Home) is added to the Laravel 5 Project. All the
resources needed by the GUI is placed in public (in case you didn't
know).

Notes:
>Authentication is being tested and will (try to) be completed in the
next coding cycle.

Change log:
>All other routes and controllers deleted.
>database.php config is changed to env settings. (.env at the root
should be edited to fit your configurations).
>Database schema name config (at .env) is changed to iteach. I would
prefer that we all use this schema for the project.
>Hash has been replaced with bcrypt() function as per convention in
Laravel 5. This is just the same but less typing.

PS:
>Please include in the SRS that the Password encryption type we use is
the bcrypt which the Laravel package primarily use.